// web/src/store/modules/activity/types.ts
export interface ActivityState {
   isLoadingAccessProject: boolean
   myProjectAccess: any[]
   isLoadingGetMyProjectAccess: boolean
   accessToMyProjects: any[]
   isLoadinggGetAccessToMyProjects: boolean
   isLoadingAccessTalent: boolean
   accessToMyProfile: any[]
   isLoadingGetAccessToMyProfile: boolean
   activities: any[]
   isLoadingActivities: boolean
   hasMore: boolean
   skip: number
   limit: number
   deleteSaveArticleActivity: any[]
   isLoadingDeleteSaveArticleActivity: boolean
   projectDetailsActivity: any[]
   isLoadingProjectDetailsActivity: boolean
}